[S07P31C202-114] 스터디 활동 현황 수정

+++ backend/connection/src/main/java/com/ssafy/connection/dto/GetDateAndCountInterface.java
@@ -4,5 +4,5 @@ import java.time.LocalDate;
 
 public interface GetDateAndCountInterface {
     LocalDate getDate();
-    Long getCount();
+    int getCount();
 }
